refactor(#32): 컨트롤러 테스트 로직 분리
