+++ src/main/java/deepdivers/community/domain/member/controller/open/MemberOpenController.java
@@ -10,17 +10,15 @@ import deepdivers.community.domain.token.dto.TokenResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/members")
 @RequiredArgsConstructor
-public class MemberOpenOpenController implements MemberOpenControllerDocs {
+public class MemberOpenController implements MemberOpenControllerDocs {
 
     private final MemberService memberService;
 
@@ -34,7 +32,7 @@ public class MemberOpenOpenController implements MemberOpenControllerDocs {
 
     @PostMapping("/login")
     public ResponseEntity<API<TokenResponse>> login(
-        @RequestBody final MemberLoginRequest request
+        @RequestBody @Valid final MemberLoginRequest request
     ) {
         final API<TokenResponse> response = memberService.login(request);
         return ResponseEntity.ok(response);
