+++ src/test/java/deepdivers/community/domain/member/service/AccountServiceTest.java
@@ -1,0 +1,7 @@
+package deepdivers.community.domain.member.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class AccountServiceTest {
+
+}
